Style the hero portion of the landing page
Edit the landing page so the style is all set, giving ample room from
the header, establishing title text, smaller text, and a sign up button,
as well as creating a placeholder for an image on the right side of
the text. These are all contained within various flex containers.